RTE-1141: Move qemu related common functions to its own file.
